# Speed up `onehotbatch`