--- a/src/onehot.jl
+++ b/src/onehot.jl
@@ -114,7 +114,7 @@ and [`onecold`](@ref) to reverse either of these, as well as to generalise `argm
 
 # Examples
 ```jldoctest
-julia> β = Flux.onehot(:b, [:a, :b, :c])
+julia> β = Flux.onehot(:b, (:a, :b, :c))
 3-element OneHotVector(::UInt32) with eltype Bool:
  ⋅
  1
@@ -163,9 +163,12 @@ If `xs` has more dimensions, `M = ndims(xs) > 1`, then the result is an
 `AbstractArray{Bool, M+1}` which is one-hot along the first dimension, 
 i.e. `result[:, k...] == onehot(xs[k...], labels)`.
 
+Notice that `xs::String` is fine, it does not need to be collected.
+For less than 32 classes, `labels::Tuple` will give best performance.
+
 # Examples
 ```jldoctest
-julia> oh = Flux.onehotbatch(collect("abracadabra"), 'a':'e', 'e')
+julia> oh = Flux.onehotbatch("abracadabra", 'a':'e', 'e')
 5×11 OneHotMatrix(::Vector{UInt32}) with eltype Bool:
  1  ⋅  ⋅  1  ⋅  1  ⋅  1  ⋅  ⋅  1
  ⋅  1  ⋅  ⋅  ⋅  ⋅  ⋅  ⋅  1  ⋅  ⋅
@@ -199,7 +202,7 @@ the same operation as `argmax(y, dims=1)` but sometimes a different return type.
 julia> Flux.onecold([false, true, false])
 2
 
-julia> Flux.onecold([0.3, 0.2, 0.5], [:a, :b, :c])
+julia> Flux.onecold([0.3, 0.2, 0.5], (:a, :b, :c))
 :c
 
 julia> Flux.onecold([ 1  0  0  1  0  1  0  1  0  0  1