Closes #1844. Before:
```julia
julia> @btime ohe_custom($sequence);  # from #1844
  min 908.784 ns, mean 962.810 ns (5 allocations, 464 bytes)

julia> @btime ohe_flux($sequence);
  min 67.083 μs, mean 69.835 μs (374 allocations, 17.30 KiB)

julia> seq = String(rand('a':'z', 10));

julia> @btime Flux.onehotbatch($seq, 'a':'z');
  min 14.750 μs, mean 15.071 μs (84 allocations, 3.84 KiB)
```
After:
```julia
julia> @btime Flux.onehotbatch($sequence, $bases_dna);  # vector of labels
  min 857.246 ns, mean 920.303 ns (4 allocations, 528 bytes)

julia> @btime Flux.onehotbatch($sequence, $(Tuple(bases_dna)));  # faster if you make the tuple
  min 586.572 ns, mean 639.020 ns (3 allocations, 496 bytes)

julia> seq = String(rand('a':'z', 10));

julia> @btime Flux.onehotbatch($seq, 'a':'z');  # 26 classes, quite a long tuple
  min 1.883 μs, mean 1.946 μs (5 allocations, 480 bytes)

julia> @btime Flux.onehotbatch($seq, $(Tuple('a':'z')));
  min 717.391 ns, mean 774.093 ns (3 allocations, 208 bytes)
```
There is probably more that could be done, e.g. to make `batch` type-stable & save a copy, but this is a start.